+++ src/components/group/LimitGroupCard.tsx
@@ -1,4 +1,20 @@
-import LimitGroupBtn from "./LimitGroupBtn";
+import { Button } from "../ui/button";
+
+const LimitGroupBtn: React.FC = () => {
+  return (
+    <Button
+      className="className= 
+        flex flex-col justify-center w-40 h-12
+        bg-blue-950 text-white
+        rounded cursor-pointer"
+      asChild
+    >
+      <a href={`${import.meta.env.VITE_PRAY_KAKAO_CHANNEL_CHAT_URL}`}>
+        문의하기
+      </a>
+    </Button>
+  );
+};
 
 export const LimitGroupCard = () => {
   const maxGroupCount = Number(import.meta.env.VITE_MAX_GROUP_COUNT);
